netconfig.py is ready for beta testing!
netconfig.py is ready for beta testing! Currently, the only way to use it is to import it and write a completely new program to leverage its functionalities. The net netconfig.py contains major changes including a separate class for network data handling (import, export, and display to user) called NetworkDataHandler. New files are created for experimentations.